+++ src/audio/google_hotword_detect.c
@@ -6,6 +6,7 @@
 
 #include <sof/audio/buffer.h>
 #include <sof/audio/component.h>
+#include <sof/audio/data_blob.h>
 #include <sof/audio/format.h>
 #include <sof/audio/kpb.h>
 #include <sof/audio/ipc-config.h>
